Develop

+++ src/recruiter/components/recruitment-phase-card.component.vue
@@ -1,12 +1,15 @@
 <template>
   <div :class="[
       'max-h-container',
-      'border',
       'rounded-xl',
       'px-4',
-      'py-4',
-      'w-72',
+      'py-2',
+      'w-full',
       'space-y-2',
+      'hover:cursor-pointer',
+      'hover:scale-95',
+      'hover:border-primary',
+      'duration-200',
       bgColor,
       bgColorDark
       ]"
@@ -16,7 +19,7 @@
         <h1 class="font-semibold text-lg">{{ phase.title }}</h1>
         <pv-button
             @click="phaseDialog = true"
-            icon="pi pi-ellipsis-h" text rounded size="large" severity="secondary" outlined
+            icon="pi pi-ellipsis-h" text rounded size="large" severity="secondary"
             aria-label="Phase dialog expander" />
         <pv-dialog v-model:visible="phaseDialog" modal :header=phase.title class="w-1/2">
           <div class="grid gap-2">
@@ -42,8 +45,8 @@
                       type="text"
                   />
                 </div>
-                <pv-button outlined type="submit"
-                           :enabled="!this.loading"
+                <pv-button type="submit"
+                           :disabled="this.loading"
                            :label="this.loading ? $t('loading') : $t('accept')"
                 />
               </div>
@@ -51,7 +54,7 @@
             <div class="grid w-full" v-if="isFirstPhase===false">
               <pv-button type="submit"
                          severity="danger"
-                         :enabled="!this.loading"
+                         :disabled="this.loading"
                          :label="this.loading ? $t('loading') : $t('delete')"
                          @click="deletePhase(phase)"
               />
@@ -59,31 +62,37 @@
           </div>
         </pv-dialog>
       </div>
-      <div>
-        <h1 v-if="candidates.length === 0">{{ $t('recruitment.no-candidates-found-message')}}</h1>
-      </div>
-      <div v-for="candidate in phase.candidates" :key="candidate.id">
-        <div @click="openCandidateDialog(candidate.id)"
-             class="flex border px-2 py-4 rounded-lg bg-white hover:scale-95 duration-100 items-center space-x-2">
-          <img class="rounded-full w-12 h-12 object-cover"
-               :src="candidate.user.profileImageUrl"
-               :alt="candidate.user.name + ' profile picture'"/>
-          <p>{{ candidate.user.name }}</p>
-        </div>
-      </div>
-      <pv-dialog v-model:visible="candidateDialog" modal>
-        <p>
-          {{ currentCandidate.user.name }}
-          <br>
-          Building...
-        </p>
-        <template #footer>
-          <pv-button label="Okay" icon="pi pi-times" @click="candidateDialog = false" text />
-        </template>
-      </pv-dialog>
+<!--      <div>-->
+<!--        <h1 v-if="this.applications.length === 0">{{ $t('no-applications-found-message')}}</h1>-->
+<!--      </div>-->
+<!--      <div v-for="application in applications" :key="application.id">-->
+<!--        <div @click="openCandidateDialog(application.id)"-->
+<!--             class="flex border px-2 py-4 rounded-lg bg-slate-50 dark:bg-slate-800 hover:scale-95 duration-100 hover:cursor-pointer items-center space-x-2">-->
+<!--          <img class="rounded-full w-12 h-12 object-cover"-->
+<!--               :src="application.applicant.profilePictureUrl"-->
+<!--               :alt="application.applicant.firstname + ' profile picture'"/>-->
+<!--          <div class="grid">-->
+<!--            <p class="font-medium">{{ application.applicant.firstname }} {{ application.applicant.lastname }}</p>-->
+<!--            <p class="text-xs">{{$t('application-date')}}: {{ formatDate(application.application_date) }}</p>-->
+<!--          </div>-->
+
+<!--        </div>-->
+<!--      </div>-->
+<!--      <pv-dialog v-model:visible="candidateDialog" modal>-->
+<!--        <p>-->
+<!--          {{ currentCandidate.user.name }}-->
+<!--          <br>-->
+<!--          Building...-->
+<!--        </p>-->
+<!--        <template #footer>-->
+<!--          <pv-button label="Okay" icon="pi pi-times" @click="candidateDialog = false" text />-->
+<!--        </template>-->
+<!--      </pv-dialog>-->
     </div>
     <div v-else>
-      <pv-spinner />
+      <div class="flex items-center justify-center">
+        <pv-spinner />
+      </div>
     </div>
   </div>
 </template>
@@ -110,18 +119,21 @@ export default {
     isFirstPhase: {
       type: Boolean,
       default: false
-    }
+    },
+    applications : {
+      type: Array,
+      default: []
+    },
   },
   data() {
     return {
       phaseDialog: false,
       candidateDialog: false,
       currentCandidate: {},
-      candidates: [],
       title:'',
       description: '',
       loading: false,
-      phaseService: new RecruitmentPhaseApiService()
+      phaseService: new RecruitmentPhaseApiService(),
     }
   },
   created() {
@@ -129,12 +141,6 @@ export default {
     this.description = this.phase.description;
   },
   methods: {
-    openCandidateDialog(candidateId) {
-      this.currentCandidate = this.phase.candidates.find(
-          candidate => candidate.id === candidateId
-      );
-      this.candidateDialog = true;
-    },
     updatePhase() {
       this.loading = true;
 
@@ -163,10 +169,13 @@ export default {
             this.$toast.add({
               severity: "success",
               summary: "JobSync",
-              detail: "Created",
+              detail: this.$t('edited'),
               life: 2000
             });
-            this.$emit('update');
+            this.phase.title = this.title;
+            this.phase.description = this.description;
+            this.phaseDialog = false;
+            this.candidateDialog = false;
           })
           .catch((error) => {
             this.$toast.add({
@@ -175,7 +184,6 @@ export default {
               detail: "There was an error. Please try again later: " + error.message,
               life: 2000
             });
-            console.log(error)
           });
 
       this.loading = false;
@@ -186,6 +194,18 @@ export default {
       if (this.isFirstPhase === true) {
         return;
       }
+
+      if (this.applications.length > 0) {
+        this.loading = false;
+        this.$toast.add({
+          severity: "error",
+          summary: "JobSync",
+          detail: this.$t('delete-phase-warn'),
+          life: 2000
+        });
+        return;
+      }
+
 
       this.phaseService.deletePhase(this.phase.id)
           .then(() => {
@@ -207,7 +227,7 @@ export default {
             console.log(error)
           });
       this.loading = false;
-    }
+    },
   }
 }
 </script>
